[CS-70] Fix Postgres update script
* Restart instead of reload system services
* Add notif_mail to provbase table in schema import
* Remove PHP Interface that is causing error due to wrong definitions
* Fix grafana user not present when no grafana was installed

+++ Install/before_upgrade.sh
@@ -1,5 +1,9 @@
 # TODO: Remove this whole code after NMSPrime v3.2
 # The whole code just does the initial steps that need to be done before Laravel can be used again with Pgsql
+# source environment variables to use php 8
+source /etc/profile.d/modules.sh
+module load php80
+
 #
 # Postgresql
 #
@@ -7,7 +11,7 @@
 systemctl enable postgresql-13.service
 systemctl start postgresql-13.service
 
-if [ ! "$(sudo -u postgres /usr/pgsql-13/bin/psql -XtAc "SELECT 1 FROM pg_database WHERE datname='DB_NAME'" )" = '1' ]; then
+if [ ! "$(sudo -u postgres /usr/pgsql-13/bin/psql -XtAc "SELECT 1 FROM pg_database WHERE datname='nmsprime'" )" = '1' ]; then
     sudo -u postgres /usr/pgsql-13/bin/psql -c 'create database nmsprime;'
 fi
 
@@ -25,28 +29,62 @@ fi
 read -r -a auths <<< $(grep '^DB_DATABASE\|^DB_USERNAME\|^DB_PASSWORD' /etc/nmsprime/env/global.env | sort | cut -d'=' -f2 | xargs)
 
 # Avoid pgloader exceptions
-mysql -u "${auths[2]}" --password="${auths[1]}" "${auths[0]}" --exec="
-    UPDATE configfile set device = 'cm' where device is NULL or device = '';
-    UPDATE contract set birthday = NULL where birthday = '0000-00-00';
-    UPDATE item set valid_from = NULL where valid_from = '0000-00-00';
-    UPDATE item set valid_to = NULL where valid_to = '0000-00-00';
-    UPDATE mta set type = 'sip' where type is NULL or type = '';
-    UPDATE oid set type = 'u' where type = '' or type is null;
-    UPDATE phonetariff set voip_protocol = 'SIP' where voip_protocol = '' or voip_protocol is null;
-    UPDATE sepamandate set valid_to = NULL where valid_to = '0000-00-00';
-    ALTER TABLE billingbase modify rcd smallint(6) null;
-    ALTER TABLE billingbase modify cdr_offset smallint(6) null;
-    ALTER TABLE debt modify indicator smallint(6) null;
-    ALTER TABLE item modify payed_month smallint(6) null;
-    ALTER TABLE settlementrun modify month smallint(6) null;
-    ALTER TABLE invoice modify month smallint(6) null;
-    ALTER TABLE costcenter modify billing_month smallint(6) null;
-    ALTER TABLE mpr drop column type;
-    ALTER TABLE mpr drop column prio;
-    ALTER TABLE mprgeopos drop column name;
-    ALTER TABLE mprgeopos drop column description;
-    ALTER TABLE netelement drop column id_name;
-"
+laravelModules=$(php /var/www/nmsprime/artisan module:list | cut -d'|' -f2)
+
+if echo "$laravelModules" | grep -q "ProvBase"; then
+    mysql -u "${auths[2]}" --password="${auths[1]}" "${auths[0]}" --exec="
+        UPDATE configfile set device = 'cm' where device is NULL or device = '';
+        UPDATE contract set birthday = NULL where birthday = '0000-00-00';
+    "
+fi
+
+if echo "$laravelModules" | grep -q "BillingBase"; then
+    mysql -u "${auths[2]}" --password="${auths[1]}" "${auths[0]}" --exec="
+        UPDATE item set valid_from = NULL where valid_from = '0000-00-00';
+        UPDATE item set valid_to = NULL where valid_to = '0000-00-00';
+        UPDATE sepamandate set valid_to = NULL where valid_to = '0000-00-00';
+        ALTER TABLE billingbase modify rcd smallint(6) null;
+        ALTER TABLE billingbase modify cdr_offset smallint(6) null;
+        ALTER TABLE item modify payed_month smallint(6) null;
+        ALTER TABLE settlementrun modify month smallint(6) null;
+        ALTER TABLE invoice modify month smallint(6) null;
+        ALTER TABLE costcenter modify billing_month smallint(6) null;
+    "
+fi
+
+if echo "$laravelModules" | grep -q "OverdueDebts"; then
+    mysql -u "${auths[2]}" --password="${auths[1]}" "${auths[0]}" --exec="
+        ALTER TABLE debt modify indicator smallint(6) null;
+    "
+fi
+
+if echo "$laravelModules" | grep -q "ProvVoip"; then
+    mysql -u "${auths[2]}" --password="${auths[1]}" "${auths[0]}" --exec="
+        UPDATE mta set type = 'sip' where type is NULL or type = '';
+        UPDATE phonetariff set voip_protocol = 'SIP' where voip_protocol = '' or voip_protocol is null;
+    "
+fi
+
+if echo "$laravelModules" | grep -q "HfcReq"; then
+    mysql -u "${auths[2]}" --password="${auths[1]}" "${auths[0]}" --exec="
+        ALTER TABLE netelement drop column id_name;
+    "
+fi
+
+if echo "$laravelModules" | grep -q "HfcCustomer"; then
+    mysql -u "${auths[2]}" --password="${auths[1]}" "${auths[0]}" --exec="
+        ALTER TABLE mpr drop column type;
+        ALTER TABLE mpr drop column prio;
+        ALTER TABLE mprgeopos drop column name;
+        ALTER TABLE mprgeopos drop column description;
+    "
+fi
+
+if echo "$laravelModules" | grep -q "HfcSnmp"; then
+    mysql -u "${auths[2]}" --password="${auths[1]}" "${auths[0]}" --exec="
+        UPDATE oid set type = 'u' where type = '' or type is null;
+    "
+fi
 
 read -r -a credentials <<< $(grep '^ROOT_DB_USERNAME\|^ROOT_DB_PASSWORD=' /etc/nmsprime/env/root.env | cut -d '=' -f2)
 mysql -u "${credentials[0]}" -p"${credentials[1]}" --exec='Create user psqlconverter; GRANT select ON *.* TO psqlconverter;'
@@ -55,27 +93,4 @@ sudo -u postgres /usr/pgsql-13/bin/psql -d nmsprime -c "
     CREATE USER ${auths[2]} PASSWORD '${auths[1]}';
 "
 
-sudo -u postgres psql nmsprime < /etc/nmsprime/sql-schemas/nmsprime.pgsql
-
-echo "LOAD DATABASE
-  FROM mysql://psqlconverter@localhost/nmsprime
-  INTO postgresql:///nmsprime
-  WITH data only, truncate, batch rows = 5000, prefetch rows = 5000
-  EXCLUDING TABLE NAMES MATCHING 'hfcbase','nas','radacct','radcheck','radgroupcheck','radgroupreply','radippool','radpostauth','radreply','radusergroup'
-    ;" > /tmp/nmsprime.load
-
-sudo -u postgres pgloader -q /tmp/nmsprime.load
-sudo -u postgres pgloader mysql://psqlconverter@localhost/nmsprime postgresql:///nmsprime
-
-sudo -u postgres /usr/pgsql-13/bin/psql -d nmsprime -c "
-    GRANT USAGE, CREATE ON SCHEMA ${auths[0]} TO ${auths[2]};
-    GRANT ALL PRIVILEGES ON ALL Tables in schema ${auths[0]} TO ${auths[2]};
-    GRANT ALL PRIVILEGES ON ALL SEQUENCES IN SCHEMA ${auths[0]} TO ${auths[2]};
-    GRANT SELECT ON ALL TABLES IN SCHEMA ${auths[0]} TO grafana;
-    GRANT USAGE ON SCHEMA ${auths[0]} TO grafana;
-"
-
-sed -i 's/^#RADIUS_DB/RADIUS_DB/' /etc/nmsprime/env/provbase.env
-sed -i "s/^RADIUS_DB_PASSWORD=.*$/RADIUS_DB_PASSWORD=$(pwgen 12 1)/" /etc/nmsprime/env/provbase.env
-
-/opt/remi/php80/root/usr/bin/php /var/www/nmsprime/artisan config:cache
+php /var/www/nmsprime/artisan config:cache
